Fix overflow bug in conv_btc_to_satoshi()

+++ core/src/conv.c
@@ -1,3 +1,5 @@
 #include "conv.h"
 
-uint64_t conv_btc_to_satoshi(uint32_t btc) { return btc * 100000000; }
+uint64_t conv_btc_to_satoshi(uint32_t btc) {
+    return (uint64_t) btc * 100000000ull;
+}
